stock_trader cleaned up. matrix class continues to be worked on. The matrix class will use an in-place inversion method that's similar to LU decomposition so that it can have O(2/3(n**3)) time complexity than the current one that uses the adjugate, and the determinant. That version takes O(2/3(n**3)) for the determinant, then we have to operate on each element of the matrix via the cofactor. So in the end it's really. A much worse time complexity of O(((n-1)**2)*(n**2)) for just the adjugate. Then also have to multiply this adjugate matrix times the determinant. Which will take O(n**2) time.
So our final time for the old method is O((n**2)+(((n-1)**2)*(n**2))).
This compared to LUDcomposition which is just a measily O(2/3*(n**3))
shows a massive speedup. In real times for a 4x4 matrix old method would
take 160 operations whereas this one will just take a measley ~43
operations. Almost 1/4 the amount of operations which should really come
in handy in the end.

This method is still going to be used for modular inverses until I can
find a better way of doing it. But all other inversions will use
LUDecomposition.

Or to put it in concrete numbers. To calculate

+++ itp232/stock_trader.h
@@ -111,9 +111,9 @@ class Inventory{
 		* when I declare the account class.
 		*/
 		//purchase an item.
-		int purchase_item(Account &usr,Item &itm, int total_items);
+		int purchase_item(Account &usr,Item &itm, unsigned int total_items);
 		//sell the item.
-		int sell_item(Account &usr, Item &itm, int total_items);
+		int sell_item(Account &usr, Item &itm, unsigned int total_items);
 		//sell all of the items.
 		void sell_all_items(Account &usr);
 		
@@ -169,7 +169,8 @@ class Inventory{
 		}
 		
 		std::cout << std::endl;
-	}		
+	}
+
 };
 //Account class.
 class Account{
@@ -219,8 +220,8 @@ class Account{
 		
 		//friend functions fully defined as a method of another class not available
 		//to anything else.
-		friend int Inventory::purchase_item(Account &usr,Item &item,int total_items);
-		friend int Inventory::sell_item(Account &usr,Item &item,int total_items);
+		friend int Inventory::purchase_item(Account &usr,Item &item,unsigned int total_items);
+		friend int Inventory::sell_item(Account &usr,Item &item,unsigned int total_items);
 		friend void Inventory::sell_all_items(Account &usr);
 };
 
@@ -232,7 +233,7 @@ class Account{
 * Item &item is an Item pointer to the current item.
 * returns a status code.
 */
-int Inventory::purchase_item(Account &usr,Item &item,int total_items){
+int Inventory::purchase_item(Account &usr,Item &item,unsigned int total_items){
 	int amount=0;
 	//get their current balance.
 	double available_funds = usr.get_balance();
@@ -240,7 +241,7 @@ int Inventory::purchase_item(Account &usr,Item &item,int total_items){
 	double total_cost = 0;
 	double price=item.get_price();
 	//figure out how many they can buy.
-	int can_buy = floor(available_funds / price);
+	double can_buy = floor(available_funds / price);
 	//get ready to print our stuff.
 	move_and_clear_terminal(4,total_items);
 	//show the menu to them.
@@ -281,7 +282,7 @@ int Inventory::purchase_item(Account &usr,Item &item,int total_items){
 * Returns:
 	Int < Status code> 0 == Nothing happened. 1 == success. -1 == failure.
 */
-int Inventory::sell_item(Account &usr, Item &item,int total_items){
+int Inventory::sell_item(Account &usr, Item &item,unsigned int total_items){
 	unsigned int amount =0;
 	//figure out how many they own.
 	unsigned int owned=total_owned[item.get_id()];
